+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 - A new generic `read` method on `UVCal` that, like the `read` methods on our other
 objects, supports all file types and a new `from_file` class method to allow one-line
 reading of UVCal files.
+- A new `UVData.new()` method (based on new function `new_uvdata`) that creates a new,
+self-consisten `UVData` object from scratch from a set of flexible input parameters.
 
 ### Deprecated
 - Reading in multiple files to `UVCal` using file-type specific read methods